Add CSS and HTML. Make the game playable within HTML
Add CSS and HTML to the page to add UI and make the game playable within the browser
instead of resorting to the console.

Some functionalities have been compromised like rounds selector and custom username.
Also the logic of the program has changed to hard code check for 5 rounds and iterate within
the global scope instead of the gameWrapper function. This ensures the desired behavior
while compromising certain reusability and adaptability.